test(e2e): make no-isolate module-sharing fixtures order-independent
The `sharing` fixtures assumed test files execute in alphabetical order
(mockShareFirst before mockShareSecond, surfaceThird after surfaceSecond),
but the runner does not guarantee file execution order. When the glob
returned files in a different order, the guards flaked on macOS.

Rework the fixtures so each guard holds regardless of scheduling:
- replace mockShareFirst/Second with symmetric peers mockShareA/mockShareB;
  whichever runs second proves clearMocks resets the shared-module mock
  across the file boundary.
- replace surfaceFirst/Second/Third with two peers surfaceA/surfaceB driving
  a shared runSurfaceGuard(); an afterAll-count invariant proves the shared
  afterAll fired for each file without assuming which ran first.

No product code change.
