+++ e2e/no-isolate/fixtures/sharing/surfaceA.test.ts
@@ -1,0 +1,4 @@
+import { runSurfaceGuard } from './surfaceHelper';
+
+// Peer of surfaceB — see surfaceHelper.runSurfaceGuard for the shared rationale.
+runSurfaceGuard('surfaceA');
