+++ e2e/no-isolate/fixtures/sharing/surfaceB.test.ts
@@ -1,0 +1,4 @@
+import { runSurfaceGuard } from './surfaceHelper';
+
+// Peer of surfaceA — see surfaceHelper.runSurfaceGuard for the shared rationale.
+runSurfaceGuard('surfaceB');
