--- e2e/no-isolate/fixtures/sharing/surfaceFirst.test.ts
@@ -1,6 +1,0 @@
-import { test } from './surfaceHelper';
-
-// Exists only so surfaceSecond is a NON-FIRST file in the reused worker.
-test('surfaceFirst: warmup', ({ expect }) => {
-  expect(true).toBe(true);
-});
